# Speed up interface load time if a layout is already selected