Ref. https://github.com/Fliplet/fliplet-studio/issues/5131

A `setupLayoutSelector()` function is created to set up the layout selector, which only does anything if a layout isn't already selected. This is previously being executed (and layout selector added to the DOM) even if the UI never gets used. This loads 28.4MBs of GIF into the browser and slows down the overall time to render.

Locally, this has reduced the time to render from ~8.5s to ~4.5s.

On staging, the time to render is currently ~3s.
On production, the time to render is currently ~7s.