@tonytlwu 2020-02-05
**Update** The production render time has reduced from ~7s to ~3s.